fix so that player's spawn in a corner of the map corresponding to their join order.
codex prompt:
===========
Two issues need to be fixed:

All 4 players need to be added to each game scene. Presently, the client can only see their player. Remark that line 74 of GameScene only has a single player in the game, whereas this should contain all players in the game.
The player's position should depend on what order they joined the queue. In particular, the spawn parameter of line 79 should use the spawn parameter of the PlayerSpawnOptions